+++ core/image_accessor.h
@@ -8,31 +8,83 @@
 
 OIDN_NAMESPACE_BEGIN
 
-  template<typename T>
   struct ImageAccessor
   {
     char* ptr;
-    size_t wByteStride; // pixel stride in number of bytes
     size_t hByteStride; // row stride in number of bytes
-    int W, H;           // width, height
+    size_t wByteStride; // pixel stride in number of bytes
+    DataType dataType;  // data type
+    int C, H, W;        // channels (1-3), height, width
 
     OIDN_HOST_DEVICE_INLINE size_t getByteOffset(int h, int w) const
     {
       return size_t(h) * hByteStride + size_t(w) * wByteStride;
     }
 
+    template<typename T = float>
     OIDN_HOST_DEVICE_INLINE vec3<T> get3(int h, int w) const
     {
-      T* pixel = reinterpret_cast<T*>(ptr + getByteOffset(h, w));
-      return vec3<T>(pixel[0], pixel[1], pixel[2]);
+      const void* pixelPtr = ptr + getByteOffset(h, w);
+      if (dataType == DataType::Float32)
+      {
+        const float* pixel = static_cast<const float*>(pixelPtr);
+        if (C == 3)
+          return vec3<T>(pixel[0], pixel[1], pixel[2]);
+        else if (C == 2)
+          return vec3<T>(pixel[0], pixel[1], pixel[1]);
+        else // if (C == 1)
+          return vec3<T>(pixel[0], pixel[0], pixel[0]);
+      }
+      else // if (dataType == DataType::Float16)
+      {
+        const half* pixel = static_cast<const half*>(pixelPtr);
+        if (C == 3)
+          return vec3<T>(pixel[0], pixel[1], pixel[2]);
+        else if (C == 2)
+          return vec3<T>(pixel[0], pixel[1], pixel[1]);
+        else // if (C == 1)
+          return vec3<T>(pixel[0], pixel[0], pixel[0]);
+      }
     }
 
+    template<typename T>
     OIDN_HOST_DEVICE_INLINE void set3(int h, int w, const vec3<T>& value) const
     {
-      T* pixel = reinterpret_cast<T*>(ptr + getByteOffset(h, w));
-      pixel[0] = value.x;
-      pixel[1] = value.y;
-      pixel[2] = value.z;
+      void* pixelPtr = ptr + getByteOffset(h, w);
+      if (dataType == DataType::Float32)
+      {
+        float* pixel = static_cast<float*>(pixelPtr);
+        if (C == 3)
+        {
+          pixel[0] = value.x;
+          pixel[1] = value.y;
+          pixel[2] = value.z;
+        }
+        else if (C == 2)
+        {
+          pixel[0] = value.x;
+          pixel[1] = value.y;
+        }
+        else // if (C == 1)
+          pixel[0] = value.x;
+      }
+      else // if (dataType == DataType::Float16)
+      {
+        half* pixel = static_cast<half*>(pixelPtr);
+        if (C == 3)
+        {
+          pixel[0] = value.x;
+          pixel[1] = value.y;
+          pixel[2] = value.z;
+        }
+        else if (C == 2)
+        {
+          pixel[0] = value.x;
+          pixel[1] = value.y;
+        }
+        else // if (C == 1)
+          pixel[0] = value.x;
+      }
     }
   };