+++ core/unet_filter.cpp
@@ -332,16 +332,24 @@ OIDN_NAMESPACE_BEGIN
     if (!output)
       throw Exception(Error::InvalidOperation, "output image not specified");
 
-    if (((color  && color->getFormat()  != Format::Float3) ||
-         (albedo && albedo->getFormat() != Format::Float3) ||
-         (normal && normal->getFormat() != Format::Float3)) &&
-        ((color  && color->getFormat()  != Format::Half3) ||
-         (albedo && albedo->getFormat() != Format::Half3) ||
-         (normal && normal->getFormat() != Format::Half3)))
+    auto isSupportedFormat = [](Format format)
+    {
+      return format == Format::Float3 || format == Format::Half3 ||
+             format == Format::Float2 || format == Format::Half2 ||
+             format == Format::Float  || format == Format::Half;
+    };
+
+    if ((color  && !isSupportedFormat(color->getFormat()))  ||
+        (albedo && !isSupportedFormat(albedo->getFormat())) ||
+        (normal && !isSupportedFormat(normal->getFormat())))
       throw Exception(Error::InvalidOperation, "unsupported input image format");
 
-    if (output->getFormat() != Format::Float3 && output->getFormat() != Format::Half3)
+    if (!isSupportedFormat(output->getFormat()))
       throw Exception(Error::InvalidOperation, "unsupported output image format");
+
+    Image* input = color ? color.get() : (albedo ? albedo.get() : normal.get());
+    if (input->getC() != output->getC())
+      throw Exception(Error::InvalidOperation, "input/output image channel count mismatch");
 
     if ((color  && (color->getW()  != output->getW() || color->getH()  != output->getH())) ||
         (albedo && (albedo->getW() != output->getW() || albedo->getH() != output->getH())) ||
@@ -427,9 +435,9 @@ OIDN_NAMESPACE_BEGIN
 
     // Get the number of input channels
     int inputC = 0;
-    if (color)  inputC += color->getC();
-    if (albedo) inputC += albedo->getC();
-    if (normal) inputC += normal->getC();
+    if (color)  inputC += 3; // always broadcast to 3 channels
+    if (albedo) inputC += 3;
+    if (normal) inputC += 3;
 
     // Create global operations (not part of any model instance or graph)
     std::shared_ptr<Autoexposure> autoexposure;
@@ -441,9 +449,9 @@ OIDN_NAMESPACE_BEGIN
     size_t totalMemoryByteSize = 0;
 
     // Create model instances for each engine of the device
-    for (int instanceId = 0; instanceId < device->getNumEngines(); ++instanceId)
+    for (int instanceID = 0; instanceID < device->getNumEngines(); ++instanceID)
     {
-      auto& instance = instances[instanceId];
+      auto& instance = instances[instanceID];
       auto& graph = instance.graph;
 
       // Create the model graph
@@ -490,20 +498,20 @@ OIDN_NAMESPACE_BEGIN
       size_t scratchByteSize = graphScratchByteSize;
 
       // Allocate scratch for global operations
-      if (instanceId == 0 && hdr)
+      if (instanceID == 0 && hdr)
         scratchByteSize = max(scratchByteSize, autoexposure->getScratchAlignedSize());
 
       // If doing in-place _tiled_ filtering, allocate a temporary output image
       ImageDesc outputTempDesc(output->getFormat(), W, H);
       size_t outputTempByteOffset = SIZE_MAX;
-      if (instanceId == 0 && inplace && (tileCountH * tileCountW) > 1)
+      if (instanceID == 0 && inplace && (tileCountH * tileCountW) > 1)
       {
         outputTempByteOffset = scratchByteSize;
         scratchByteSize += outputTempDesc.getAlignedSize();
       }
 
       // Check the total memory usage
-      if (instanceId == 0)
+      if (instanceID == 0)
       {
         totalMemoryByteSize = scratchByteSize + graph->getPrivateByteSize() +
                               (graphScratchByteSize + graph->getPrivateByteSize()) * (device->getNumEngines() - 1);
@@ -515,18 +523,18 @@ OIDN_NAMESPACE_BEGIN
       }
 
       // Allocate the scratch buffer
-      auto scratch = device->getEngine(instanceId)->newScratchBuffer(scratchByteSize);
+      auto scratch = device->getEngine(instanceID)->newScratchBuffer(scratchByteSize);
 
       // Set the scratch buffer for the graph and the global operations
       graph->setScratch(scratch);
-      if (instanceId == 0 && hdr)
+      if (instanceID == 0 && hdr)
         autoexposure->setScratch(scratch);
 
       // Finalize the network
       graph->finalize();
 
       // Create the temporary output image
-      if (instanceId == 0 && outputTempByteOffset < SIZE_MAX)
+      if (instanceID == 0 && outputTempByteOffset < SIZE_MAX)
         outputTemp = scratch->newImage(outputTempDesc, outputTempByteOffset);
 
       instance.inputProcess  = inputProcess;