+++ devices/sycl/sycl_engine.h
@@ -63,6 +63,16 @@ OIDN_NAMESPACE_BEGIN
         [=](sycl::nd_item<N> it) { f(it); });
     }
 
+    // Enqueues a work-group kernel with explicit subgroup size
+    template<int subgroupSize, int N, typename F>
+    OIDN_INLINE void submitKernel(WorkDim<N> numGroups, WorkDim<N> groupSize, const F& f)
+    {
+      lastEvent = syclQueue.parallel_for<F>(
+        sycl::nd_range<N>(numGroups * groupSize, groupSize),
+        getDepEvents(),
+        [=](sycl::nd_item<N> it) [[intel::reqd_sub_group_size(subgroupSize)]] { f(it); });
+    }
+
     // Enqueues a basic ESIMD kernel
     template<int N, typename F>
     OIDN_INLINE void submitESIMDKernel(WorkDim<N> globalSize, const F& f)